Main refactored
