+++ CMakeLists.txt
@@ -98,6 +98,7 @@ add_library(TetrisEngineCore STATIC
     src/Engine.cpp
     src/NeuralNetwork.cpp
     src/Game.cpp
+    src/UI.cpp
 )
 
 target_include_directories(TetrisEngineCore PUBLIC
