+++ src/main.cpp
@@ -4,38 +4,12 @@
 #include <onnxruntime_cxx_api.h>
 #include "../include/TetrisEngine/Board.h"
 #include "../include/TetrisEngine/Game.h"
+#include "../include/TetrisEngine/Ui.h"
 #include <raylib.h>
 #include <imgui.h>
 #include <rlImGui.h>
 
 using namespace tetris;
-
-std::string PieceTypeToString(PieceType pt) {
-    switch (pt) {
-        case PieceType::I: return "I";
-        case PieceType::O: return "O";
-        case PieceType::T: return "T";
-        case PieceType::S: return "S";
-        case PieceType::Z: return "Z";
-        case PieceType::J: return "J";
-        case PieceType::L: return "L";
-        default: return " ";
-    }
-}
-
-Color GetColorForPiece(PieceType pt) {
-    static std::unordered_map<PieceType, Color> colorMap = {
-        {PieceType::EMPTY, DARKGRAY},
-        {PieceType::I, SKYBLUE},
-        {PieceType::O, YELLOW},
-        {PieceType::T, PURPLE},
-        {PieceType::S, GREEN},
-        {PieceType::Z, RED},
-        {PieceType::J, BLUE},
-        {PieceType::L, ORANGE},
-    };
-    return colorMap.count(pt) ? colorMap[pt] : RAYWHITE;
-}
 
 int main() {
     // All this does is test to make sure onnxruntime is importing properly.
@@ -76,146 +50,18 @@ int main() {
     while (!WindowShouldClose()) {
         BeginDrawing();
         ClearBackground(BLACK);
-
         rlImGuiBegin();
-        // offset to prevent overlap from squares
 
-        // this is box for controlling without keybinds
-        ImGui::SetNextWindowPos(ImVec2(600, 100));
-        ImGui::Begin("Controls");
+        gameOver = DrawControlsPanel(game, commandHistory, gameOver);
 
-        if (!gameOver) {
-            // these controls were also reversed
-            if (ImGui::Button("Left") || IsKeyPressed(KEY_LEFT)) {
-                game.getBoard(0).MoveActivePiece(-1, 0);
-                commandHistory.push_back("Move Left");
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("Right") || IsKeyPressed(KEY_RIGHT)){ 
-                game.getBoard(0).MoveActivePiece(1, 0);
-                commandHistory.push_back("Move Right");
-            }
-            
-            if (ImGui::Button("Soft Drop") || IsKeyPressed(KEY_DOWN)) {
-                commandHistory.push_back("Drop 1");
-                if (!game.getBoard(0).MoveActivePiece(0, -1)) {
-                    game.getBoard(0).LockActivePiece();
-
-                    if (!game.getBoard(0).SpawnRandomPiece()) gameOver = true;
-                    game.getBoard(0).GetNextQueue();
-
-                   
-                }
-            }
-            
-            ImGui::SameLine();  
-            if (ImGui::Button("Hard Drop") || IsKeyPressed(KEY_SPACE)) {
-                game.getBoard(0).HardDropActivePiece();
-                commandHistory.push_back("Drop X");
-                if (!game.getBoard(0).SpawnRandomPiece()) gameOver = true;
-                game.getBoard(0).GetNextQueue();
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("HOLD") || IsKeyPressed(KEY_LEFT_SHIFT)) {
-                game.getBoard(0).HoldPiece();
-                commandHistory.push_back("Hold Piece");
-            }
-
-            if (ImGui::Button("Rotate CW") || IsKeyPressed(KEY_E)){ 
-                game.getBoard(0).RotateActivePiece(RotationDirection::CLOCKWISE);
-                commandHistory.push_back("Rotate CW");
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("Rotate CCW") || IsKeyPressed(KEY_Q)) {
-                game.getBoard(0).RotateActivePiece(RotationDirection::COUNTER_CLOCKWISE);
-                commandHistory.push_back("Rotate CCW");
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("Rotate 180") || IsKeyPressed(KEY_W)) {
-                game.getBoard(0).RotateActivePiece(RotationDirection::ONE_EIGHTY);
-                commandHistory.push_back("Rotate 180");
-            }
-
-            if (ImGui::Button("Reset") || IsKeyPressed(KEY_T)) {
-                game.getBoard(0).Reset();
-                commandHistory.push_back("Reset Board");
-            }
-           
-           // erase first command if bigger than 10
-            if (commandHistory.size() > 10) {
-                commandHistory.erase(commandHistory.begin());
-            }
-        }
-
-        if (gameOver) {
-            ImGui::Text("Game Over!");
-        }    
-        
-        ImGui::Text("Score: %d", game.getBoard(0).GetScore());
-        ImGui::Text("Lines: %d", game.getBoard(0).GetLinesCleared());
-        ImGui::End();
-
-        ImGui::SetNextWindowPos(ImVec2(450, 100));
-        ImGui::SetNextWindowSize(ImVec2(100, 0));
-        ImGui::Begin("Queue");
-        for (PieceType pt : game.getBoard(0).GetNextQueue()) {
-            Color c = GetColorForPiece(pt);
-            ImGui::ColorButton("##color", ImVec4{c.r / 255.0f, c.g / 255.0f, c.b / 255.0f, 1.0f},
-                       ImGuiColorEditFlags_NoTooltip | ImGuiColorEditFlags_NoDragDrop,
-                       ImVec2(30, 30));
-            ImGui::SameLine();
-            ImGui::TextUnformatted(PieceTypeToString(pt).c_str());
-        }
-        ImGui::End();
-
-        ImGui::SetNextWindowPos(ImVec2(450, 350));
-        ImGui::SetNextWindowSize(ImVec2(100, 0));
-        ImGui::Begin("Hold");
-        
-        PieceType held = game.getBoard(0).GetHeldPieceType();
-        if (held != PieceType::EMPTY) {
-            Color c = GetColorForPiece(held);
-            ImGui::ColorButton("##heldcolor", 
-                ImVec4{c.r/255.0f, c.g/255.0f, c.b/255.0f, 1.0f},
-                ImGuiColorEditFlags_NoTooltip | ImGuiColorEditFlags_NoDragDrop,
-                ImVec2(30, 30));
-            ImGui::SameLine();
-            ImGui::TextUnformatted(PieceTypeToString(held).c_str());
-        } else {
-            ImGui::Text("Empty");
-        }
-        ImGui::End();
-
-        // this is the box to see command history
-        ImGui::SetNextWindowPos(ImVec2(600, 350));
-        ImGui::Begin("Command History");
-        // populate
-        ImGui::BeginChild("ScrollRegion", ImVec2(250, 200), true, 0);
-        for (const std::string& cmd : commandHistory) {
-            ImGui::TextUnformatted(cmd.c_str());
-        }
-        ImGui::EndChild();
+        DrawQueuePanel(game.getBoard(0));
+        DrawHoldPanel (game.getBoard(0));
+        DrawHistoryPanel(commandHistory);
 
         rlImGuiEnd();
 
-        // Draw game.getBoard(0)
-        std::vector<PieceType> state(VISIBLE_BOARD_HEIGHT * BOARD_WIDTH, PieceType::EMPTY);
-        state = game.getBoard(0).GetRenderableState();
-
-        for (int row = 0; row < VISIBLE_BOARD_HEIGHT; ++row) {
-            for (int col = 0; col < BOARD_WIDTH; ++col) {
-                PieceType pt = state[row * BOARD_WIDTH + col];
-                Color color = GetColorForPiece(pt);
-
-                DrawRectangle(
-                    boardOffsetX + col * cellSize,
-                    boardOffsetY + row * cellSize,
-                    cellSize - 2,
-                    cellSize - 2,
-                    color
-                );
-            }
-        }
+        // draw the actual tetris board
+        DrawBoardGrid(game.getBoard(0), boardOffsetX, boardOffsetY, cellSize);
 
         EndDrawing();
     }
